# academify/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Academify API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the Student Community & Collaboration Platform — COMP6703001"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://e2526-wads-b4ac-02.csbihub.id",
      "description": "Production server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "HTTP-only session cookie set by Better Auth after sign-in. Sign in via the app or POST /api/auth/sign-in/email, then use Try it out from the same origin."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Legacy alias — Academify uses session cookies, not Bearer JWT."
      }
    },
...
  },
  "paths": {
...
    "/api/search": {
      "get": {
        "summary": "Global search (users, forums, threads)",
        "tags": [
          "Search"
        ],
        "security": [
          {
            "sessionCookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 10,
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results"
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      }
    },
...
    "/api/admin/users": {
      "get": {
        "summary": "Get all users (Admin only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "sessionCookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string",
              "enum": [
                "student",
                "instructor",
                "moderator",
                "admin"
              ]
            },
            "description": "Filter by role"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "suspended",
                "banned"
              ]
            },
            "description": "Filter by account status"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search by name or email"
          }
        ],
        "responses": {
          "200": {
            "description": "List of users"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden - Admin only"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/comments/{commentId}": {
      "put": {
        "summary": "Update a comment",
        "tags": [
          "Comments"
        ],
        "security": [
          {
            "sessionCookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "commentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content"
                ],
                "properties": {
                  "content": {
                    "type": "string",
                    "example": "Updated comment content."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Comment updated successfully"
          },
          "400": {
            "description": "Missing required fields"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Comment not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete a comment",
        "tags": [
          "Comments"
        ],
        "security": [
          {
            "sessionCookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "commentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comment deleted successfully"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Comment not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/moderation/delete/{postId}": {
      "post": {
        "summary": "Delete a post via moderation (Moderator/Admin only)",
        "tags": [
          "Moderation"
        ],
        "security": [
          {
            "sessionCookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "example": "Violates community guidelines."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post deleted successfully"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden - Moderator/Admin only"
          },
          "404": {
            "description": "Post not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}